copy mbr.bin
For making a bootable usb drive, we need the MBR in usb drive works
correctly. This mbr.bin is used to dd to the usb drive. Actually
livecd-iso-to-disk.sh would do this for you, but that script copy the
mbr.bin from local system, so if you run usb.sh in ubuntu to build a
centOS usb drive, it must fail. As a result, the best way is copy
mbr.bin here, and dd again to make our CentOS usb drive bootable.